third_party/WebKit/LayoutTests/platform/win7/external/wpt/web-animations/timing-model/timelines/document-timelines-expected.txt

This is a testharness.js-based test.
FAIL Document timelines report current time relative to navigationStart assert_equals: The current time matches requestAnimationFrame time expected 16.733000054955482 but got 24.108666460961103
Harness: the test ran to completion.
